+++ cocos/3d/physics/util.ts
@@ -1,3 +1,4 @@
+import { Vec3 } from '../../core/value-types';
 import { quat, vec3 } from '../../core/vmath';
 
 export function stringfyVec3 (value: {x: number; y: number; z: number}): string {
@@ -14,4 +15,20 @@ export function stringfyQuat (value: {x: number; y: number; z: number; w: number
     } else {
         return `(x: ${value.x}, y: ${value.y}, z: ${value.z}, w: ${value.w})`;
     }
+}
+
+interface IWrapped<T> {
+    __cc_wrapper__: T;
+}
+
+export function setWrap<Wrapper> (object: any, wrapper: Wrapper) {
+    (object as IWrapped<Wrapper>).__cc_wrapper__ = wrapper;
+}
+
+export function getWrap<Wrapper> (object: any) {
+    return (object as IWrapped<Wrapper>).__cc_wrapper__;
+}
+
+export function maxComponent (v: Vec3) {
+    return Math.max(v.x, Math.max(v.y, v.z));
 }
